racingcar (수정)
- 1단계
- 학습테스트 실습
